+++ Sources/StreamChat/Config/ChatClientConfig.swift
@@ -190,6 +190,9 @@ public struct ChatClientConfig {
     /// The `URLSessionConfiguration` being used as default configuration for the `APIClient` and
     /// `WebSocketClient`
     public var urlSessionConfiguration: URLSessionConfiguration = .default
+    
+    /// How many hours the unsent messages should be queued for sending when the internet connection is available.
+    public var unsentMessagesRetryHoursThreshold: CGFloat = 12
 
     public init(
         apiKey: APIKey